moved user page into correct folder
